refactor: separate trackDuration function into utils
