+++ src/components/Search/SearchResult/TabContent/TrackTab/TrackItem.tsx
@@ -1,7 +1,8 @@
 import { Loading } from '@/components/Main';
 import { PLACEHOLDER_IMAGE } from '@/constants/image';
 import { Track } from '@/models/Track';
-import { useEffect, useState } from 'react';
+import { UTC2HHMMSS } from '@/utils';
+import { useEffect, useMemo, useState } from 'react';
 import Card from '../_shared/Card';
 import CoverImage from '../_shared/CoverImage';
 import CoverImageSkeleton from '../_shared/CoverImage/CoverImageSkeleton';
@@ -28,19 +29,7 @@ const TrackItem = ({
     }
   }, [artistInfo]);
 
-  /**
-   * 밀리초로 되어 있는 트랙 재생 시간을 HH:MM:SS 형식으로 변환
-   */
-  const trackDurationToHHMMSS = (() => {
-    const time = new Date(item.duration_ms);
-    const MMSS = [
-      `${time.getUTCMinutes()}`.padStart(2, '0'),
-      `${time.getUTCSeconds()}`.padStart(2, '0'),
-    ];
-    return time.getUTCHours() > 0
-      ? [`${time.getUTCHours()}`.padStart(2, '0'), ...MMSS].join(':')
-      : MMSS.join(':');
-  })();
+  const trackDurationHHMMSS = useMemo(() => UTC2HHMMSS({ item }), [item]);
 
   return (
     <Card
@@ -63,7 +52,7 @@ const TrackItem = ({
       }
       isPlayable={item?.is_playable}
     >
-      <li>재생 시간: {trackDurationToHHMMSS}</li>
+      <li>재생 시간: {trackDurationHHMMSS}</li>
       {validTrackArtistInfo ? (
         <MultiProfile artist={artistInfo} />
       ) : (
