+++ src/utils/index.ts
@@ -1,3 +1,5 @@
+import { Track } from '@/models/track';
+
 /**
  * 참고 : https://gist.github.com/ca0v/73a31f57b397606c9813472f7493a940
  * */
@@ -10,4 +12,18 @@ export function debounce<F extends (...args: Parameters<F>) => ReturnType<F>>(
     clearTimeout(timeout);
     timeout = setTimeout(() => func(...args), waitFor);
   };
+}
+
+/**
+ * 밀리초로 되어 있는 UTC 시간을 HH:MM:SS 형식으로 변환
+ */
+export function UTC2HHMMSS({ item }: { item: Track }) {
+  const time = new Date(item.duration_ms);
+  const MMSS = [
+    `${time.getUTCMinutes()}`.padStart(2, '0'),
+    `${time.getUTCSeconds()}`.padStart(2, '0'),
+  ];
+  return time.getUTCHours() > 0
+    ? [`${time.getUTCHours()}`.padStart(2, '0'), ...MMSS].join(':')
+    : MMSS.join(':');
 }
